# bucket: allow to allocate key on stack in Put()