@cenkalti 2023-08-09
Cool! Can we have a short comment in code that explains why `newKey` var is introduced instead of reusing `key`?