@fyfyrchik 2023-08-10
Fixed, please, take a look.
BTW, #532 is also affected -- we might want to fix it there or here depending on the PR merge order.