@ahrtr 2023-08-10
> BTW, #532 is also supposedly affected -- we might want to fix it there or here depending on the PR merge order.

The PR https://github.com/etcd-io/bbolt/pull/532 has already been merged. Please rebase this PR and fix it as well. thx